remove replaceMany, improve replace
