+++ browser.js
@@ -854,65 +854,64 @@ extendInterfaces('stub', function(context, teardown) {
 });
 
 /**
- * Stubs Polymer.Base.instanceTemplate with function that calls
- * provided `domDecorator` function before returning the dom.
- *
- * Sets up teardown to restore the stub.
+ * Maintains a collection of DOM decorator functions.
+ * Applies them by stubbing Polymer.Base.instanceTemplate
  */
-function stubInstanceTemplate(domDecorator, teardown) {
+function DomStub() {
+  this._decorators = [];
+};
+
+/**
+ * Returns true, if this instance wasn't setup yet (or just teared down)
+ */
+DomStub.prototype.pristine = function() {
+  return this._decorators.length === 0;
+};
+
+/**
+ * Add decorator function
+ */
+DomStub.prototype.decorator = function(decorator) {
+  this._decorators.push(decorator);
+};
+
+/**
+ * Removes all setup decorators and restores instanceTemplate stub
+ */
+DomStub.prototype.teardown = function() {
+  this._decorators = [];
+
+  // Restore the stubbed version of `Polymer.Base.instanceTemplate`:
+  if (Polymer.Base.instanceTemplate.isSinonProxy) {
+    Polymer.Base.instanceTemplate.restore();
+  }
+};
+
+/**
+ * Stubs Polymer.Base.instanceTemplate with function that applies
+ * provided decorator to the dom.
+ */
+DomStub.prototype.setup = function() {
   // Keep a reference to the original `Polymer.Base.instanceTemplate`
   // implementation for later:
   var originalInstanceTemplate = Polymer.Base.instanceTemplate;
+
+  var decorators = this._decorators;
 
   // Use Sinon to stub `Polymer.Base.instanceTemplate`:
   sinon.stub(Polymer.Base, 'instanceTemplate', function(template) {
     // The DOM to replace is the result of calling the "original"
     // `instanceTemplate` implementation:
     var dom = originalInstanceTemplate.apply(this, arguments);
-    domDecorator(dom);
+
+    // Apply all decorators that were setup
+    decorators.forEach(function(decorator) {
+      decorator(dom);
+    });
 
     return dom;
   });
-
-  // After each test...
-  teardown(function() {
-    // Restore the stubbed version of `Polymer.Base.instanceTemplate`:
-    if (Polymer.Base.instanceTemplate.isSinonProxy) {
-      Polymer.Base.instanceTemplate.restore();
-    }
-  });
-}
-
-/**
- * Returns function that takes dom as a parameter and replaces
- * all oldTagNames with tagNames.
- *
- * All annotations and attributes will be set on the placement element the way
- * they were set for the original element.
- */
-function tagReplacer(oldTagName, tagName) {
-  return function(dom) {
-    // The nodes to replace are queried from the DOM chunk:
-    var nodes = Array.prototype.slice.call(dom.querySelectorAll(oldTagName));
-
-    // For all of the nodes we want to place...
-    nodes.forEach(function(node) {
-
-      // Create a replacement:
-      var replacement = document.createElement(tagName);
-
-      // For all attributes in the original node..
-      for (var index = 0; index < node.attributes.length; ++index) {
-        // Set that attribute on the replacement:
-        replacement.setAttribute(
-          node.attributes[index].name, node.attributes[index].value);
-      }
-
-      // Replace the original node with the replacement node:
-      node.parentNode.replaceChild(replacement, node);
-    });
-  };
-}
+};
 
 /**
  * replace
@@ -929,38 +928,42 @@ function tagReplacer(oldTagName, tagName) {
  * they were set for the original element.
  */
 extendInterfaces('replace', function(context, teardown) {
+
+  var stub = new DomStub();
+
   return function replace(oldTagName) {
     return {
       with: function(tagName) {
-        var decorator = tagReplacer(oldTagName, tagName);
-        return stubInstanceTemplate(decorator, teardown);
-      }
-    };
-  };
-});
 
-/**
- * replaceMany
- *
- * The replaceMany addon works just like replace addon,
- * except that you can specify multiple replacements.
- *
- * beforeEach(function() {
- *   replaceMany({
- *     'x-foo': 'x-fake-foo',
- *     'y-bar': 'y-fake-bar'
- *   });
- * });
- */
-extendInterfaces('replaceMany', function(context, teardown) {
-  return function replaceMany(tags) {
-    var decorator = function(dom) {
-      for (var oldTagName in tags) {
-        var tagName = tags[oldTagName];
-        tagReplacer(oldTagName, tagName)(dom);
+        // Stub Polymer.Base.instanceTemplate
+        if (stub.pristine()) {
+          stub.setup();
+          teardown(stub.teardown.bind(stub));
+        }
+
+        stub.decorator(function(dom) {
+          // The nodes to replace are queried from the DOM chunk:
+          var nodes = Array.prototype.slice.call(dom.querySelectorAll(oldTagName));
+
+          // For all of the nodes we want to place...
+          nodes.forEach(function(node) {
+
+            // Create a replacement:
+            var replacement = document.createElement(tagName);
+
+            // For all attributes in the original node..
+            for (var index = 0; index < node.attributes.length; ++index) {
+              // Set that attribute on the replacement:
+              replacement.setAttribute(
+                node.attributes[index].name, node.attributes[index].value);
+            }
+
+            // Replace the original node with the replacement node:
+            node.parentNode.replaceChild(replacement, node);
+          });
+        });
       }
     };
-    return stubInstanceTemplate(decorator, teardown);
   };
 });
