+++ browser/mocha.js
@@ -12,7 +12,6 @@ import * as config from './config.js';
 import './mocha/fixture.js';
 import './mocha/stub.js';
 import './mocha/replace.js';
-import './mocha/replace-many.js';
 import { applyExtensions } from './mocha/extend.js';
 
 // Mocha global helpers, broken out by testing method.
